reduce size of loop
